Consodlidate utils into one file
