+++ client/src/features/user/components/Tabs/Account.tsx
@@ -1,7 +1,7 @@
 import { useField } from "@/hooks/useField";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { changeEmail, changeName } from "../../api/supabase";
-import { notify } from "@/utils/notify";
+import { notify } from "@/utils";
 import { AuthError } from "@supabase/supabase-js";
 
 export const Account = () => {
