--- client/src/utils/notify.ts
@@ -1,5 +1,0 @@
-import { useNotificationStore } from "../stores/notificationStore";
-
-export const notify = (type: "info" | "error" | "success", message: string) => {
-  useNotificationStore.setState({ type, message });
-};
