--- client/src/utils/sortAsc.ts
@@ -1,6 +1,0 @@
-import { Plant } from "@/features/calendar/types";
-
-export const sortAsc = (plants: Plant[]) =>
-  plants.sort((a, b) =>
-    a.name.localeCompare(b.name, "en", { sensitivity: "base" })
-  );
